+++ src/token_index.rs
@@ -4,15 +4,15 @@ use crate::cache::ReserveCache;
 
 #[derive(Debug)]
 pub struct TokenIndexMap {
-    pub address_to_index: HashMap<H160, u16>,
-    pub index_to_address: HashMap<u16, H160>,
+    pub address_to_index: HashMap<H160, u32>,
+    pub index_to_address: HashMap<u32, H160>,
 }
 
 impl TokenIndexMap {
     pub fn build_from_reserve_cache(reserve_cache: &ReserveCache) -> Self {
         let mut address_to_index = HashMap::new();
         let mut index_to_address = HashMap::new();
-        let mut next_index: u16 = 0;
+        let mut next_index: u32 = 0;
 
         for entry in reserve_cache.iter() {
             let token0 = entry.value().token0;
